[REF] refactor file_autotask_rel
