+++ file_autotask_rel/__init__.py
@@ -21,4 +21,4 @@
 #
 ###############################################################################
 
-import file_buffer
+import file_document
