--- file_buffer/file_buffer_data.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0"?>
-<openerp>
-    <data noupdate="1">
-
-        <record forcecreate="True" id="ir_cron_run_file_buffer_scheduler_action" model="ir.cron">
-            <field name="name">File Buffer: process</field>
-            <field eval="False" name="active"/>
-            <field name="user_id" ref="base.user_root"/>
-            <field name="interval_number">10</field>
-            <field name="interval_type">minutes</field>
-            <field name="numbercall">-1</field>
-            <field eval="False" name="doall"/>
-            <field eval="'file.buffer'" name="model"/>
-            <field eval="'run_file_buffer_scheduler'" name="function"/>
-            <field eval="'()'" name="args"/>
-        </record>
-
-        <record forcecreate="True" id="ir_cron_check_state_file_buffer_scheduler_action" model="ir.cron">
-            <field name="name">File Buffer: check state (output buffer)</field>
-            <field eval="False" name="active"/>
-            <field name="user_id" ref="base.user_root"/>
-            <field name="interval_number">10</field>
-            <field name="interval_type">minutes</field>
-            <field name="numbercall">-1</field>
-            <field eval="False" name="doall"/>
-            <field eval="'file.buffer'" name="model"/>
-            <field eval="'check_state_file_buffer_scheduler'" name="function"/>
-            <field eval="'()'" name="args"/>
-        </record>
-
-    </data>
-</openerp>
