--- file_buffer/file_buffer_view.xml
@@ -1,106 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  file_buffer for OpenERP
-  Copyright (C) 2012 Akretion David BEAL <david.beal@akretion.com>
-  The licence is in the file __openerp__.py -->
-
-<openerp>
-    <data>
-        <!-- SEARCH -->
-        <record id="file_buffer_view_search" model="ir.ui.view">
-            <field name="name">file.buffer.search</field>
-            <field name="model">file.buffer</field>
-            <field name="type">search</field>
-            <field name="arch" type="xml">
-                <search string="Search Files Buffer">
-                    <group>
-                        <filter icon="kanban-pause" name="wait" string="Waiting"
-                                domain="[('state','=','waiting')]"/>
-                        <filter icon="gtk-apply" string="Done"
-                                domain="[('state','=','done')]"/>
-                        <filter icon="gtk-execute" string="Running"
-                                domain="[('state','=','running')]"/>
-                        <filter icon="terp-gtk-stop" string="Fail"
-                                domain="[('state','=','fail')]"/>
-                    </group>
-                    <group>
-                        <filter icon="gtk-go-down" string="In"
-                                domain="[('direction','=','input')]"/>
-                        <filter icon="gtk-go-up" string="Out"
-                                domain="[('direction','=','output')]"/>
-                        <field name="name" />
-                        <field name="file_id" string="Exchange file"/>
-                    </group>
-                    <field name="name" />
-                    <field name="file_id" string="Exchange file"/>
-                    <field name="state"/>
-               </search>
-            </field>
-        </record>
-
-        <!-- FORM -->
-        <record id="file_buffer_view_form" model="ir.ui.view">
-            <field name="name">file.buffer.view_form</field>
-            <field name="model">file.buffer</field>
-            <field name="type">form</field>
-            <field name="arch" type="xml">
-                <form string="File Buffer">
-                    <field name="name" />
-                    <field name="state" readonly="1"/>
-                    <field name="file_id" />
-                    <field name="job_ended" />
-                    <field name="active"/>
-                    <field name="direction" readonly="1" />
-                    <field name="type"/>
-                    <newline />
-                    <button name="run" string="Run" colspan="2" type="object"
-                            attrs="{'invisible':[('state','in', ('running','done'))]}"/>
-                    <newline />
-                    <button name="check_state" string="Check state" colspan="2"
-                            type="object" attrs="{'invisible':[('state','!=', 'running')]}"/>
-                    <newline />
-                    <separator string="External application response (for output direction)" colspan="4"/>
-                    <field name="response" nolabel="1" readonly="1" colspan="4" />
-                </form>
-            </field>
-        </record>
-
-        <!-- TREE -->
-        <record id="file_buffer_view_tree" model="ir.ui.view">
-            <field name="name">file.buffer.view_tree</field>
-            <field name="model">file.buffer</field>
-            <field name="type">tree</field>
-            <field name="arch" type="xml">
-                <tree string="File Buffer">
-                    <field name="name" />
-                    <field name="state" />
-                    <field name="active" string="act." />
-                    <field name="file_id" />
-                    <field name="job_ended" />
-                    <field name="direction" string="direct." />
-                    <button name="run" string="Run"
-                            icon="gtk-go-forward" type="object"/>
-                </tree>
-            </field>
-        </record>
-
-        <!-- Action used by menu -->
-        <record id="action_file_buffer" model="ir.actions.act_window">
-            <field name="name">File Buffer</field>
-            <field name="res_model">file.buffer</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-            <field name="context">{"search_default_wait":1}</field>
-        </record>
-
-        <menuitem name="File Exchange"
-                  id="menu_file_exchange_root"
-                  parent="base.menu_administration"/>
-
-        <menuitem id="menu_file_buffer"
-                  action="action_file_buffer"
-                  parent="menu_file_exchange_root"
-                  sequence="20" />
-
-    </data>
-</openerp>
